pyproject.toml
[tool.poetry]
name = "fastapi-auth-template"
version = "0.1.0"
description = ""
authors = ["draincoder <draincoder@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.10"
alembic = "^1.10.3"
fastapi = "^0.95.1"
python-dateutil = "^2.8.2"
pyyaml = "^6.0"
uvicorn = "^0.21.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
colorlog = "^6.7.0"
asyncpg = "^0.27.0"
python-multipart = "^0.0.6"
orjson = "^3.8.10"
opentelemetry-instrumentation-fastapi = "^0.38b0"
opentelemetry-instrumentation-logging = "^0.38b0"
opentelemetry-exporter-otlp = "^1.17.0"
opentelemetry-api = "^1.17.0"
opentelemetry-sdk = "^1.17.0"
prometheus-client = "^0.16.0"
bcrypt = "^4.0.1"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.2.2"
pytest-asyncio = "^0.20.3"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
